export/requirements.txt
flask
flask_cors
psycopg2-binary
pyjwt
requests
